fix: move commands_internal back to commands
