+++ src-tauri/src/airdrop.rs
@@ -21,8 +21,10 @@
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 use jsonwebtoken::{decode, Algorithm, DecodingKey, Validation};
-use log::warn;
+use log::{info, warn};
 use serde::{Deserialize, Serialize};
+
+use crate::{mm_proxy_adapter::MergeMiningProxyConfig, UniverseAppState};
 
 const LOG_TARGET: &str = "tari::universe::airdrop";
 
@@ -86,4 +88,31 @@ pub async fn validate_jwt(airdrop_access_token: Option<String>) -> Option<String
             None
         }
     })
+}
+
+pub async fn restart_mm_proxy_with_new_telemetry_id(
+    state: tauri::State<'_, UniverseAppState>,
+) -> Result<(), String> {
+    info!(target: LOG_TARGET, "Restarting mm_proxy");
+    let telemetry_id = state
+        .telemetry_manager
+        .read()
+        .await
+        .get_unique_string()
+        .await;
+    let mm_proxy_manager_config = state
+        .mm_proxy_manager
+        .config()
+        .await
+        .ok_or("mm proxy config could not be found")?;
+    let _unused = state
+        .mm_proxy_manager
+        .change_config(MergeMiningProxyConfig {
+            coinbase_extra: telemetry_id.clone(),
+            ..mm_proxy_manager_config
+        })
+        .await
+        .map_err(|e| e.to_string());
+    info!(target: LOG_TARGET, "mm_proxy restarted");
+    Ok(())
 }
